Moving app files to app module
