+++ src/modules/app/__tests__/authReducer-test.js
@@ -1,9 +1,9 @@
 import configureMockStore from 'redux-mock-store';
-import createMiddleware from '../../redux/clientMiddleware';
-import ApiClient from '../../redux/ApiClient';
+import createMiddleware from '../../../redux/clientMiddleware';
+import ApiClient from '../../../redux/ApiClient';
 import request from 'superagent';
 import nocker from 'superagent-nock';
-import Config from '../../config';
+import Config from '../../../config';
 
 import * as actions from '../appReducer';
 import auth from '../appReducer';
